fix: cold start fonts

+++ notion4ever/site_generation.py
@@ -124,7 +124,9 @@ def generate_site(structured_notion: dict, config: dict):
     generate_css(config)
     logging.debug("🤖 SASS translated to CSS folder.")
 
-    shutil.rmtree(Path(config["output_dir"]) / "css" / "fonts")
+
+    if (Path(config["output_dir"]) / "css" / "fonts").exists():
+        shutil.rmtree(Path(config["output_dir"]) / "css" / "fonts")
     shutil.copytree(Path(config["sass_dir"]) / "fonts", Path(config["output_dir"]) / "css" / "fonts")
     logging.debug("🤖 Copied fonts.")
